+++ lib/rails_admin/config/fields/association.rb
@@ -57,6 +57,11 @@ module RailsAdmin
           @associated_collection_cache_all ||= (associated_model_config.abstract_model.count < 100)
         end
 
+        # Reader whether association's elements can be removed
+        def removable
+          association.foreign_key_nullable?
+        end
+
         # Reader for the association's child model's configuration
         def associated_model_config
           @associated_model_config ||= RailsAdmin.config(association.klass)